[MPP-25599] Allow building with OS versions of libblas, liblapack, libf2c
Previously, they had to be present in the ext-directory. With this change,
the build will pick them up from anywhere in the system, but still prefer
the versions in 'ext', if present. This makes it easier to build GPDB
outside the normal build environment.

We should clean up the whole libgpla thing; the functions that need those
libraries should probably be put into an extension that's loaded
dynamically, and that extension could then link with those extra libraries
as usual. And of course, we should not link statically with them in the
first place. But this is a good first step.