[MPP-25599] Build ORCA glue code using same flags as the rest of the tree.
I don't see any reason for being different. This code is linked directly
to the backend like any other code; the ORCA library could be built with
other flags but that's a different story.

(This is part of an effort to remove references to releng.mk. I'm grepping
for references to the variables defined in releng.mk)