[MPP-25599] Revert 38eba909, which removed GPOS_DEBUG flag from gpopt build.
I didn't realize that the GPOS_DEBUG flag is affected, but the buildfarm
pointed out..

If the glue code is compiled with the GPOS_DEBUG flag, you must link with
the debug-enabled version of libgpos. There are some inline functions in
the headers that reference the fEnableExtendedAssert variable, if built
with GPOS_DEBUG. I think we should fix that, but for now, let's make the
buildfarm happy by reverting.