[MPP-25599] Don't use unportable -l:<library filename> linker option.
That syntax was used to force the linker to link some libraries statically,
but the buildfarm says it's not supported at least on some OS X systems.
Might be a GNUism, not sure.

Just use normal -lfoo syntax instead. In our standard build environment,
we only have static versions of these libraries in ext, so this has no
effect there. But if you build without ext (which doesn't work yet, but
that's the goal of these build system changes), the build will now pick
up the shared libraries installed on the system, instead of linking them
statically. Arguably that's even better, so let's just do that, even
though I originally tried to avoid any unnecessary differences between
building with or without ext.