[MPP-25599] Backport --with-libxml option.
When the XML support was backported from PostgreSQL 8.3, the configure
option was left out. Bring that in, so that you can build without XML
support.

I already backported the USE_LIBXML ifdefs in earlier commit, but I didn't
realize that we're missing the configure option too, so XML support was
always disabled. Oops.