[MPP-25659] Don't call CHECK_FOR_INTERRUPTS() while sending a log message
If there's a die-signal pending, we might recurse into logging another
message. If that happens a few times, you will reach the error stack depth,
which causes a PANIC.

I'm not sure that's exactly what happened in the customer's system, but
processing a signal while sending a log message was definitely involved.

Most calls to this function happen while interrupts are held (e.g. the
EmitErrorReport call in the main query processing loop), and this doesn't
make a difference to those callers.

While we're at it, fix comments to explain what the purpose of the
retry-loop is, and fix indentation and formatting.