[MPP-25631] Fix compiler warnings about variable being used uninitialized.
We had deliberately suppressed these warnings a long time ago already, in
commit 33924da8b517fba4bafac7644f798f5b1994b159. I missed that, and
overzeolously reverted it back to the upstream version, as it seemed like
there was no need for that difference as there were no other material
changes to the tuplesort_begin_* functions. But we had removed some code
in SelectSortFunction(), and old versions of gcc compiler didn't realize
that elog(ERROR) never returns, and hence SelectSortFunction() in fact
always initializes the variables, or throws an error. I didn't get such
warnings on GCC 4.4 or later.

To fix, instead of putting the spurious variable initializations to where
they were, silence the warning SelectSortFunction(), so that the change is
localized to where the problem came from.