[MPP-25770][#87800216] Enable to run gptransfer --help with another gptransfer still running.
Currently, gives out an error when we want to run gptrasnfer --help. We
want to ensure that the user can still run --help option without having
to wait for gptransfer to finish.

[#87800216] Fix UnicodeDecodeError: 'ascii' codec can't decode

python cannot decode parts of gptransfer help and needs to be switched to
pattern matching, since the default for the systems is apparently ascii.

source cluster should be the one that runs the setup not dest cluster...

Assume that database have been created already by the setup.