[JIRA: MPP-25735] Assign the owner of pg_aovisimap table to be the same as the parent ao table.
When upgrading from 4.2 to 4.3 the upgrade script creates pg_aoseg.pg_aovisimap
tables. However the relowner of these tables is always set to gpadmin instead
of the owner of the parent ao table. This causes issues when running vacuum.
This fix correctly assigns the owner of the visimap table to be the same as its
parent ao table.