[MPP-25631] Get rid of compile-time parsing of CaQL queries.
The actual generated code was dead, #ifdef'd out. All CaQl queries are (and
were before this too) parsed and processed on first invocation, and cached
thereafter. Remove a lot of the supporting infrastructure, regression tests
etc. that is not needed anymore either.

Aside from the removing the gperf hash table, which was #ifdef'd out, this
turns the caql_getcstring_plus() and caql_getoid_plus() functions into more
traditional functions that look at the attribute's type OID, and extract
the value accordingly. They used to use a switch-statement with a case for
each catalog table in the system, which was generated by the calico.pl
script. But we have the attribute's type OID available so we might as well
use that instead.

With these changes, catquery.c is no longer auto-generated during the
build. It's now a regular C file, included in the repository as usual.