[MPP-25631] Remove JSON module.
It was used by calico.pl, but isn't needed anymore. (Including 3rd party
perl modules in random locations inside the source tree was a bad idea
from the get go...)

tidycat.pl still needs JSON, but that is run manually, not as part of the
normal build process. Install JSON on your development system to run it.