[MPP-25631] catquery.c cleanup
Remove a bunch of now-unnecessary #includes. Turn caql_heapclose() into
a function; the "if (..) { } else " with a dangling "else" macro was scary.
The usual macro pattern used in PostgreSQL to make the macro call look like
a function is:

do {
  ...
} while(0)

Also remove the unused caql_PKLOCK function, and the related fields and
code.

That pattern should be used for macros like this. But a function is nicer
from debugging point of view etc. anyway, so let's just do that.