[MPP-25631] Look up syscache ids dynamically, when parsing a CaQL query.
This isn't too exciting on its own, but removes the need to have the
syscache ids in the catcore structs, which let's us simplify the script
that generates the catcore structs.

Scanning the cacheinfo array to find the syscache is slower than having the
syscacheid already in the struct at compile time, but it shouldn't make any
difference in practice. It's only performed when the CaQL query is parsed,
i.e. the first time the CaQL query is executed, and cached thereafter. (We
could optimize scanning the cacheinfo array, which is currently a simple
linear scan, with a hash table or similar if that became an issue)