[MPP-25631] Replace the catcoregen.py python script with a perl script.
Perl is the language used in PostgreSQL for generating code like this.
Python is not required, at build-time. Greenplum requires python at runtime
anyway, for all the gpMgmt tools, so that's not so important for us. The
more important reason for this rewrite is that this replacement Perl script
reads the *.h headers directly, and does not rely on the .json file.

As far as the server build is concerned, the catalog JSON file, tidycat, and
all that infrastructure is no longer required. That makes it possible to
build the server repository on its own, without the "main" repository (the
catalog JSON files were stored in the main repository). You'll still need to
use tidycat if you actually modify the catalog schema, but that'll happpen
on your desktop at development time, and the resulting changes will be
committed to the repository, as always.