[MPP-25798] Speed up and simplify tidycat2 test case a bit.
Execute all the CREATE FUNCTION statements in a single transaction block -
that makes it a lot faster. The regression test took about 1 minute on the
buildfarm before this change, and about 1/2 minute after this. Not a huge
difference, but still worth doing.

While we're at it, simplify the way the SQL script is generated and executed
a bit. psql has a \! command for executing arbitrary shell commands, no need
to create an external table for that.