[MPP-25631] Mock test Makefile kibitzing.
Move the include of Makefile.global from Makefile.mock to the individual
Makefiles. This may seem repetetive, but Makefile.global is included in
all other Makefiles like this anyway, so let's be consistent. This gives
a chance to define things between including Makefile.global and
Makefile.mock, which might come handy.