[MPP-25631] Replace custom tidygram.py script with upstream check_keywords.pl
Both scripts do the same thing: check that kwlist.h is in sync with gram.y.
So let's use the upstream script. I backported the latest version of this
script from upstream master branch. We had the 8.3 version in src/tools
already, but it was not used for anything, so remove that. In more recent
upstream versions, the script is in src/backend/parser, not src/tools, so
put this version in src/backend/parser as well.

The upstream script is slightly more restrictive than our tidygram.py
version: it also checks that the keywords are in alphabetical order in
gram.y. That's not a correctness issue, just nitpicking by the upstream
script, but fix the order to make the check pass.

In upstream, the script had to be run manually until version 9.3. 9.3
onwards, it's invoked automatically when compiling gram.c from the gram.y
bison source file. Since we were running tidygram.py automatically, run
check_keywords.pl automatically too, 9.3-style.