[MPP-25631] Remove dead code (optserver).
The optserver helper process was #ifdef'd out in postmaster.c.

optserver was an experiment to have a separate process to run ORCA in.
Backends would send queries to the optserver process, and optserver would
run ORCA to generate a plan, and send the plan back to the original process.
But it was never taken into production.