[MPP-25631] Avoid a few catalog lookups in cdbRelSize().
All the callers have a Relation at hand, so might as well pass it to
cdbRelSize() to avoid the overhead of catalog lookups to get the schema and
relation name.

I saw the catalog lookups show up when profiling a very simple query.
It didn't account to much in wall-clock time in the end - much of the time
is spent dispatching the query to the segments and back. But there's little
downside - the code is arguably even more readable after this patch - so
might as well optimize it.