[MPP-25865] Backport upstream bugfix.
We're not particularly concerned about this bug, compared to the zillion
others that affect us that have been fixed in upstream, but the lack of
this one was causing minor merge conflicts while trying to backport a
subsequent security fix. Instead of adjusting the other patch to apply,
seems simpler to just backport this too.

commit 2ab199b354135f199a895c4ee61c521b03a4d321
Author: Tom Lane <tgl@sss.pgh.pa.us>
Date:   Wed Sep 7 17:06:33 2011 -0400

    Fix corner case bug in numeric to_char().

    Trailing-zero stripping applied by the FM specifier could strip zeroes
    to the left of the decimal point, for a format with no digit positions
    after the decimal point (such as "FM999.").

    Reported and diagnosed by Marti Raudsepp, though I didn't use his patch.