[MPP-25631] Remove unnecessary errOmitLocation(true) calls.
Omitting the location is the default, unless errcode is
ERRCODE_INTERNAL_ERROR. Remove errOmitLocation(true) from ereport()s
that specify another errcode, as they don't do anything.

The errOmitLocation() calls are problematic for merging with upstream
PostgreSQL, as many of the calls were in code we've inherited from
upstream. They would cause merge conflicts when we try to merge with
PostgreSQL again.