[MPP-25631] Fix list of fault-injector types.
Broken in commit 26228b5c. I didn't notice that there are three places
that list the fault-injector types, and the all need to be kept in sync
with each other. To avoid this kind of bugs in the future, add comments to
all three places that they need to be kept in sync.