[MPP-25983] This patch fixes two things in upgrade 1. Determine if a segment is down. 2. Failure to rollback when there are symlinks in master data directory.
If the database is not able to stop completely, then gpstop issues kill signals
to the segments in increasing order of severity, starting from SIGCONT upto
SIGABRT and finally SIGKILL. Some of these signals will cause the segment to
exit without updating its state in pg_control file.  gpstop checks if
a segment is down by sending a kill -0 whereas gpmigrator_mirror checks the
pg_control file.  Since the process is actually gone, gpstop thinks database
has been shutdown, but gpmigrator_mirror thinks otherwise based on the
pg_control.

Secondly when gpmigrator_mirror is rolling back, it tries to restore the files
in master datadirectory from a backup. If one of the directories in the master
data directory is a symlink, it fails.