[JIRA: MPP-25830] CTE query returns wrong results.
RCA: the plan skeleton generated is right, while the extParam field of
some plan nodes is wrong, for CTE related plans; during execution,
preprocess_initplans() decide whether the InitPlan should be pre-executed
on master according to the extParam related field, so some InitPlans which
should be preprocessed are dispatched to segments, leading to wrong results.

Fix is first to find the correct PlannerInfo structure for function
create_ctescan_plan, and then to disable filter push down for CTE subquery
when the filter contains Param.