Move bitmapscan bugbuster test to main suite, and clean it up.
Many of the test queries were doubled: run first within an ignore-block,
and then for real. That seems pointless, so remove the duplicates. Also,
fix typos in comments.

This hopefully fixes test failures that we've occasionally seen in the
buildfarm, where the mechanism in gpdiff.pl that masks out result set
ordering differences gets confused by the duplicated queries in the
ignore-blocks.

Conflict: There is no greenplum_schedule in this 4.3.x branch. Moved
          the test insertion into parallel_good_schedule instead.