[MPP-25960] This fix prompts the user if he wants to resume a previously failed upgrade or to restart the upgrade from the beginning.
The upgrade tool stores all the steps it executed in a state file in master
data directory. At startup it reads this state file to determine where it
crashed and starts execution from that step onwards. If the upgrade fails, it
has a rollback mechanism where it undoes all the changes and reverts the
database back to its original state. But in some cases even the rollback may
fail and end up not cleaning the state file.

In case of failed rollback when the user re-runs upgrade, it reads the state
file and tries to complete the failed rollback which is expected. But if the
user has performed the rollback manually and forgotten to clean up the state
file, upgrade would still assume that the rollback is incomplete and try to
complete it. User would expect the rollback to start from the beginning since
rollback has been completed maually. This fix lets the user to choose if he
wants to resume the upgrade or to start a fresh upgrade.