Fix gp_dump drop schema prior to create.
During a restore on a database without the related schemas (ex. --redirect
to empty database) in which these schemas are set to be dropped, the SET
search_path can be missing before the CREATE TABLE statements in the dump
file. This results in tables being restored in the default schema instead
of their expected schema.

The main issue was that the ArchiveHandle->currSchema was not being reset
when drop schema prior to create was requested in gp_dump. The
ArchiveHandle->currSchema is now properly reset. This was originally fixed
in upstream Postgres, and this commit just ports over that fix to gp_dump.