allow classpath to be absolute
Previously the classpath was relative
to $GPHOME/lib/postgresql/java/

The implication of this is that in order to reference jars outside
of this directory the classpath will look like ../../../ etc.

This patch retains that behaviour but adds the ability to use absolute
paths.

For security purposes changing the classpath can only be done by a
super user or set in postgresql.conf

Signed-off-by: Marbin Tan <mtan@pivotal.io>