Fix the typo for external table dump
The typo is from the dumping based on object types(i.e,
index/function/trigger/table/data), but currently gpdb
does not support this feature ported from postgres.

Initially planned to remove the related dead code, but
as we do the merge of the upstream postgres, I decide to
save the code to minimize the diff and effort in case
we want to support such in future.