A bunch of commits to adopt new cmockery changes by Heikki:
Clean up mock tests in src/bin with "make clean"

Avoid assertions in mock test, rather than suppress them.

Seems cleaner.

Divide backend-specific mock stuff to separate Makefile.

This way we don't need to link the whole backend into client mock programs.
We've gotten away with it so far, but it would cause clashes if one of the
client programs contained a function or global variable with same name as
in the backend. And of course it's completely unnecessary anyway.

Simplify mock test Makefiles.

There are a bunch of backend files that are never mocked, but always linked
into test program as is. List those in Makefile.mock, so that every Makefile
doesn't need to list them. Makes the makefiles simpler, and you need less
boilerplate code when adding a new test program.

Not all of these backend files were listed in the _REAL_OBJS variable of
every test, but in those tests that they were not, they were also not used,
so it doesn't matter if we use the real or mock version.

Use real memory manager for mock tests.

Some tests were already doing this, but it seems nicer to default to using
the real memory manager for all tests. An individual test can still opt-out
by listing the relevant files in <testname>_MOCK_OBJS. Some tests had had
to jump through hoops to mock the memory manager, which is no longer
necessary.

In mock tests, link real version of every backend object file by default.

This inverts the logic of the <testname>_REAL_OBJS mechanism. We now link
with the real version of each object file, *except* those listed in
<testname>_MOCK_OBJS variable.

Refactor mock test Makefiles to only mock backend files as needed.

Instead of creating a mock version of every backend object file, only
create mock files as needed by individual tests. Speeds up the build, and
is nicer anyway.