Refactored backup and restore Behave tests
This commit represents a major overhaul of the backup and restore test suite,
with the following main goals:
1) Cut down the time required to run the tests (previous runtime was ~3.5 hours
on Pivotal test machines, now it's down to around 1 hour 15 minutes).
2) Make each test scenario independent, so that the database state at the start
of a scenario does not depend on which scenarios were run before it and failures
in one scenario will not cause later scenarios to fail as a result.
3) Group commonly used steps, remove redundant scenarios and unnecessary steps
within scenarios, standardize common database and table names, and perform
similar cleanup, both to cut down on runtime and to improve readability and
comprehensibility of the suite.