Silence compiler warning in gpfdist.c
When compiling gpfdist on Ubuntu, the compile fails because of the compiler
flag to treat all warnings as errors. You get a warning when calling
freopen and not checking the result of the call.

Added a check for NULL of the redirect of stdout and stderr on init. If it
fails, it will print a message and return out of init with a -1.