Fix assertion failure if the producer of a share scan is in subplan.
The comment in apply_shareinput_xslice() said that we process top-plan
first, to ensure that we find the producer of a shared scan before its
consumers, when the top plan contains a producer for subplans. But the
situation can just as well be the other way 'round.

No new test case added, because running the existing with_clause test
with gp_cte_sharing=on trips the assertion.

Fixes github issue #287.