Refactored NetBackup-related backup and restore Behave tests
This commit represents an overhaul of the test suite relating to Veritas NetBackup
integration with the backup and restore utilities. It has the same goals and makes
similar changes to this suite that were made to the base backup and restore suite
in commit e8889fc5438dd8cb13e39ad894446964d9c99871, and also focuses on removing
any scenarios that test functionality that isn't directly related to NetBackup and
is therefore sufficiently tested in the other suite, to improve the usability of
the suite and dramatically decrease total runtime (down from ~15 hours to ~5 hours
on Pivotal test machines).