Check EOF values for AO/CO before updating aoseg table.
For AO/CO tables due to functional nature, new EOF value to be updated
must always be greater than current value. Currently, this check is
performed on persistent table update but it happens as part of Commit-Prepared.
So its super too late to take any action, even if we detect it can only PANIC
the system and already resulted in data corruption for reads as aoseg has been
updated and transaction commited. Hence moving the check while updating the
aoseg table, when such case is detected now transaction can be aborted without
impacting the system or already loaded data in table.