Fix __gp_user_tables_readable to not error out on inaccessible schemas.
Merge down from master to fix:
If there is a schema in the database that the user does not have access
to, doing "select * from gp_toolkit.__gp_user_tables_readable" throws a
permission denied error on the schema. Use the relation-OID-variant of
has_table_privilege() function, to work around that.