Add more installcheck test coverage for AO/CO tables.
Most of these test additions are inspired from Pivotal's internal
testing and needed to be added to the open source installcheck to
give the community more test coverage on AO/CO tables.  This commit
mostly adds extra coverage for indexes and partition tables.

Conflict: known_good_schedule is parallel in 5.0.x but not in 4.3.x
          so had to add the test manually in its own row