Modified gpcheckcat to display tables having missing attributes in a more readable format.
Previously, gpcheckcat identified tables with missing attributes but
displayed them in various locations in the output in a
non-standardized format. This commit summarizes all the tables with missing
attributes in one section at the end of the output in the format
"[database].[schema].[table].[segment id]".

Authors: Chris Hajas and Jamie McAtamney