Fixing test to account for mirrors and number of segments on test machine.
Fixes commit f569c1d10eff569ba8ab3801c7a7ef588eb05c5c.  The query used
in this test to identify missing attributes returned values for both the
primary and mirror segments, and has been modified to avoid doing so.
Also, if an attribute is deleted from every segment this is detected as
an error with the master, rather than an error with each segment, so the
test now only deletes one segment's attribute so it can test for master
and segment error messages.

Authors: Chris Hajas and Jamie McAtamney