Seg WAL Replication: Primary should inform the master once primary's mirror goes down
Implemented the following.

During the processing of the user COMMIT, the Backend process waits in
SyncRepWaitForLSN() for the WAL Sender process to replicate the COMMIT XLOG
record.

The WAL Sender process finds out about the mirror going down while replicating
to the mirror. As a result it calls WalSndKill().

In WalSndKill(), we update the segment state to SegmentStateFault and fault type
to be FaultTypeMirror.

We then wait for the primary to enter Check Tracking mode.

This basically happens by the FTS process finding out about the downed mirror
during FTS probe and taking the following actions.

1. Update gp_segment_configuration table to mark the downed mirror.

2. Calling gp_primary_mirror to connect to the primary and send a transition
message to the primary to transition to Change Tracking mode.

Once the WAL Sender process discovers that the primary has transitioned to the
Change Tracking mode, then it releases the Backend process to continue with the
user COMMIT. The WAL Sender process terminates after this.

Cf. Tracker Story: https://www.pivotaltracker.com/story/show/116464881