Remove dead code in outfuncs.c (#714)
The reason to remove them is because: print_variable_fields is initialized
to true, and only set to false in outfuncs_workfile_mgr_init() function.
But outfuncs_workfile_mgr_init() is never called. So we can remove all of:
print_variable_fields, outfuncs_workfile_mgr_init(), outfuncs_workfile_mgr_end(),
and all the if() tests depending on print_variable_fields.

Co-authored-by: Kuien Liu and Heikki Linnakangas

This is ported from
https://github.com/greenplum-db/gpdb/commit/87b97b94e47c099634906035d910f8a510cfb312