Implemented the following.
In applyStepForTransitionToPrimarySegmentMode() (called through
doRequestedPrimaryMirrorModeTransitions() by the postmaster to transition the
primary to Change Tracking mode), if we are using filerep (detected by
args->hostPort to be -1), initialize XLog startup and change next state to
TSDoDatabaseStartup so that all the filerep processing is bypassed and the
primary switches to Change Tracking mode.

Cf. Tracker Story: https://www.pivotaltracker.com/story/show/118862289.