[#120036053] Fix SEGV for improper typemod initialization
For some queries that contained ROW expression with data type of
char varying (with a length of 7 larger), Greenplum Database a generated
SIGSEGV error. The error was caused by improper initialization
of record types during query execution.

In PostgreSQL, we rely on ExecEvalWholeRowVar() to do this, at
execution time. In Greenplum, however, the typmods assigned to
rowtypes must be the same between the master and all segments,
so we must assign typmods at executor startup, before the plan
gets dispatched to segments.