Add check for backup directory existence to gprecoverseg
A previous commit added the capability to gprecoverseg to copy all
backup files from the mirror to the primary during a full recovery,
but assumed that the backup directory would always exist and so
gprecoverseg -F would fail if it was not present.  This commit adds
a check to ensure that gprecoverseg -F will finish successfully if
there is no backup directory.