Enhanced pipeline to tar up and upload TINC code
Before this commit, the output of the pipeline was a .zip installer
file which was uploaded. After this commit, both the .zip installer
file and a tar.gz of the tinc portion of the source code used to
create it will be uploaded.  This will be useful for downstream jobs
(such as pulse tinc regression tests) that will at some point need the
tinc source code to be passed to them.

Authors: Ashwin Agrawal and Amil Khanzada