s3ext: fix a race condition and a exception bug
== race condition

Without this fix, ChunkBuffer::read() might be waiting for the signal
from ChunkBuffer::fill(), which has already been terminated in some
certain cases.

== exception bug

When a FAIL response from KeyReader is given, an exception
will be throw from S3Service::fetchData(), and this exception
will crash the downloader threads and process (core dumped).

We replace exception with error return, and it will be handled
by the caller.

Signed-off-by: Adam Lee <ali@pivotal.io>
Signed-off-by: Kuien Liu <kliu@pivotal.io>
Signed-off-by: Haozhou Wang <hawang@pivotal.io>