Remove inaccurate icg test that injects QueryFinishPending
This test was introduced in commit 369a7b488e33dfe263446c78aaa11662ea2c633d.
It does not actually test the changes of that commit, instead it was aiming to test some other scenarios.
It seems that the test does not have a deterministic behavior.
We remove it for now, and we will investigate if the scenarios that we were aiming to test are possible and write deterministic tests.