Resync flat files from primary to mirror on startup only if they are out of sync.
Currently when the cluster starts up, all flat files from the following
directories are unconditionally copied from the primaries to the mirrors.

pg_clog
pg_distributedlog
pg_distributedxidmap
pg_multixact/members
pg_multixact/offsets
pg_subtrans

This can take hours on a large cluster (100+ nodes).  If the files are already
in sync, there is no need to transfer them.

In this commit, we compute a md5 checksum for each of the files in the above
directories and store them in a file called 'slru_checksum_file' in each
directory.  The primary then computes the checksum of the slru_checksum_file and
sends it to the mirror.

The mirror also computes the slru_checksum_file for each of the directories in
parallel, and when it receives the checksum of the slru_checksum_file from the
primary it computes the checksum of its local slru_checksum_file and compares
it.

If the two checksums match, then no files are transferred from the primary to
the mirror for that directory.  In case of a mismatch all the files in the
directory are transferred as before.  The mismatch is logged.

Authors: Shoaib Lari, Amil Khanzada, Abhijit Subramanya