After Row Triggers: disable for AO tables
Greenplum has limited trigger capabilities which do not work with the current AO
implementation. It was found that after row triggers were being executed for AO
table DMLs using planner resulting in erroneous heap reads. Instead of adding
logic to do AO segfile reads, we disable after row triggers for AO tables as
Greenplum does not really support triggers for AO anyways.

 Conflicts between 5.0_MASTER and 4.3_STABLE:
	src/backend/commands/trigger.c
      - irrelevant for 4.3_STABLE
	src/backend/executor/execMain.c
      - moved changes to src/backend/executor/execDML.c as the functions are
        located there in 4.3_STABLE instead of execMain.c like in 5.0_MASTER
	src/test/regress/output/misc.source
      - 4.3_STABLE has less tables in misc test; just needed to add 4

The original commit, 3f96c366755a7684ef6f05e33b1466c1faa44ac5, in 5.0_MASTER
includes a fix for heap tables. This was not backported in this commit because
the issue was specific to 5.0_MASTER and does not appear in 4.3_STABLE.