Link ORCA-related objects directly to backend.
This gets rid of the libdxltranslators.so library. Instead, link the
code in the gpopt directory straight into the backend binary. Much
simpler.

A pleasant side-effect of this is that the build can now also pick up the
ORCA libraries from a path specified with "./configure --with-libs=<path>".
It's no longer necessary to have OPTIMIZER and OBJDIR_DEFAULT Makefile
variables set correctly, which is nice when building outside the
gpAux/releng infrastructure. I'm hoping we can get rid of much of that
infrastructure in the near future.

(cherry picked from commit 1c1bc21935bb8b5d063582253776357f725fab04)