Change test cases to not rely on implicit int -> text cast. [#127290641]
This is partially cherry-picked from
https://github.com/greenplum-db/gpdb/commit/6f4986ecd6f80eab000cfd0746ec1089188b95cf

For expediency we scoped the cherry-pick to only the `qp_derived_table`
instead of resolving the conflict.

The original message for commit 6f4986ec was:

The int -> text cast (and many other casts to text) were downgraded
to assignment casts in the upstream, and we're about to merge that patch
soon. To prepare for that and to reduce the size of that huge merge
commit somewhat, adjust GPDB test cases to not rely on those casts that
are about to be removed.

When the casts were downgraded in the upstream, it revealed a lot of buggy
queries in applications (or put another way, it broke a lot of applications
:-) ). Many of the queries in our regression tests look bogus like that
as well, for example the "sale.dt < 10" comparison in the qp_olap_mdqa test
(sale.dt is a date). But I didn't try to change the meaning of the queries,
I just added the casts needed to make them do the same thing they used to.
I'm afraid that changing the substance of the queries would reduce our
code coverage, if the tests are reproducing some very specific scenario.