Parameterize dist destination with the AWS bucket name [#125741031]
A recent refactor actually made it harder for people to be careful when
forking pipelines. This makes it easier for people who do fork pipelines
to make sure that their artifacts on DIST won't be clobbering others'

This requires different credentials from the gpdb-ci-deployments
repository -- do a git pull there.

Signed-off-by: C.J. Jameson <cjameson@pivotal.io>