Validate persistent information in relcache entry.
We have observed that the persistent TID and serial number associated with a
relfilenode in a relation descriptor may be updated to a new value before
relfilenode is updated.  If an XLog record is written with such a relation
descriptor, then the XLog record fails to apply during recovery.

This commit adds a check to validate sanity of persistent information in a
relation descriptor by fetching persistent information using relfilenode from
gp_relation_node.

The validation is controlled by a new GUC, gp_validate_pt_info_relcache.