Remove index usage for pg_aoseg and pg_aocsseg lookups.
AO and AOCO tables do an index lookup during inserts and compaction
against the segno column. Select statements do not use this
index. This patch will make it so that we now do a heap scan on the
pg_aoseg/pg_aocsseg tables to obtain the segment file information to
remove the work on maintaining the index. This should not impact
performance as there are only 128 tuples in the pg_aoseg/pg_aocsseg
tables and relieves some possible index corruption issues that may
uncommonly appear in rare cases. Note that this index is still updated
to allow customers to downgrade GPDB versions and hence index update
issues may still appear.

Authors: Jimmy Yih, Asim Praveen, and Abhijit Subramanya

 Conflicts:
	src/backend/access/aocs/aocssegfiles.c
          - aoEntry function argument is not in 5.0
	src/backend/access/appendonly/aosegfiles.c
          - aoEntry function argument is not in 5.0