Avoid storing maxTID for persistent tables.
It's very error prone to externally maintain variable to track
maxTID for a table, need to worry about lot of cases like vacuum and all.
Persistent tables historically used this mechanism to perform certain
checks, which seem irrelevant now that we are freeing tuples in
regular heap way for persistent tables as well.