Faultinjector status includes number of times fault hit
Currently, fault injector framework doesn't convey how many times a given fault
has been hit. This is useful information to help optimize many tests which are
currently relying on sleep and select * from pg_stat_activity to figure out if
its time to trigger next step or not. Tests which need the fault to be triggered
for n number of commands and then perform some other activity, will now be able
to get this information in gpfaultinjector status message.