Allow gptransfer to work if columns are dropped and re-added before the transfer.
When a column is dropped and readded, Postgres assigns the column the
next unused ordinal number (not its original one). Previously we
explicitly compared these ordinal numbers when doing validation between
tables for gptransfer, but this caused gptransfer to fail if ordinal
numbers were not the same, even if the actual column order was the same.
We have removed the ordinal number check and now compare by the order
to fix this. Now it is not necessary for the columns to have matching
ordinal numbers. As long as the column order is the same, the transfer
can complete successfully.

Authors: Karen Huddleston and Stephen Wu