Modified partition attribute check to ignore column ordinality.
Previously gptransfer compared attribute numbers of the partition columns
to determine if two partition tables had the same structure. However, if a
column located before the partition columns was dropped and re-added on the source
system and then a dump was taken of that table and used to create a table with
the same structure on the destination system, the two tables would have differing
partition column attributes due to the order of column creation on the source
system, and the transfer would fail.

We have modified this check to compare the positions of the partition columns
within the table instead of just looking at the attribute numbers so that
the above scenario can now run successfully.

Authors: Karen Huddleston and Jamie McAtamney