Fix the calculation of CRC.
The old crc32cFinish() function did not actually modify the
crc variable but just returned the modified value to the
caller. The FIN_CRC32() macro brought in as part of 6c025b5
modifies the crc variable in place which causes a mismatch when
comparing the crc generated using the old function. This commit
fixes the issue by making the new FIN_CRC32() macro an explicit
no-op.

Authors: Shoaib Lari and Abhijit Subramanya