Remove crashrecovery test to reindex template1 database.
The test fails intermittently due to a design flaw in crash recovery - the need
to initialize relcache fully for updating persistent tables.  This issue will be
fixed in master as part of the effort to eradicate persistent tables.

For more details, refer to gpdb-dev discussion:
https://groups.google.com/a/greenplum.org/forum/#!topic/gpdb-dev/SIUiF5FmRTg
and tracker story: https://www.pivotaltracker.com/story/show/133285039