Use signed URLs for SLES S3 artifacts when triggering pulse [#135633239]
A change to the pulse-build-scripts helped reveal that we had never
actually started using signed URLs for SLES artifacts to enable Pulse
Triggering.

The change to the pulse-build-scripts seeks to enable using the `aws
s3api` for fetching artifacts. However, it still has some limitations.
Until it is improved, the safest thing for us to do is to continue
always using signed URLs for all artifacts which Pulse fetches. This
ensures that Pulse will use `wget` in downloading the artifact, and
the `aws s3api` code path won't be used yet (it only tries to download
S3 artifact URLs that are unsigned).

Signed-off-by: C.J. Jameson <cjameson@pivotal.io>