Use non-pinned java 7 version [#136047535]
This fixes the error: alternatives --set java /usr/lib/jvm/java-1.7.0-openjdk-1.7.0.111-2.6.7.2.el7_2.x86_64/jre/bin/java

We found that the Centos:7 image from Dockerhub dropped openjdk7 from
default installation. We've added a yum install to ensure that we get a
java 7 installed.

We decided to allow the version of java 7 to float -- we had
pinned it to a specific version previously.

Signed-off-by: Jingyi Mei <jmei@pivotal.io>

[ci skip]