Fix a bug that Rollup causes query to crash
In rollup() processing, when the grouping cols are processed, the target list is
traversed for each grouping col to find a match. Each node in the target list is
mutated in `replace_grouping_columns_mutator()` to find match for grouping cols.

In the failing query:

```
SELECT COUNT(DISTINCT c) AS r
FROM (
    SELECT c
    , CASE WHEN (e = 2) THEN 1 END AS f
    , 1 AS g FROM foo
    ) f_view
GROUP BY ROLLUP(f,g);
```

grouping cols ‘g’ is a constant 1. when target entry for ‘f’ is being mutated,
we mutate all the fields of caseexpr recursively. since the `result` field in
caseexpr is a const with same value as that of ‘g’, it gets considered as a
match and we replace the node by null which later results in crash.

Essentially, since we traverse all the fields of expr, a const
node (appearing anywhere in caseexpr) with same value as that of const grouping
col will be matched.

Hence the following query also fails where one of the arg of equality operator
is a const with same value as that of ‘g’.

```
SELECT COUNT(DISTINCT c) AS r
FROM (
    SELECT c
    , CASE WHEN (e = 1) THEN 2 END AS f
    , 1 AS g FROM foo
    ) f_view
GROUP BY ROLLUP(f,g);
```

Fix:
1. refactored the `replace_grouping_columns_mutator()` to have separate
routines for target list replacement and qual replacement.
2. for target list, just look at top level node. do not recurse.
3. for quals, the top level is always some kind of expression, so we
need to recurse to find the match. Skip replacing constants here.

signed-off-by: haisheng yuan <hyuan@pivotal.io>
signed-off-by: xin zhang <xzhang@pivotal.io>