Added a flag to identify if analyze statement is being executed.
During planning tuple description is derived for every result of an
operator. However, this adds the overhead of building the cache /
looking up the cache to derived the description. In case of a analyze,
a severe performance impact is observed as it triggers a bunch of
queries.

This fix will skip the population of tuple description if ANALYZE is
running.