Remove fail-fast check for pid file.
-- previously, within a 60-second loop to test if a pqPing was
successful, there was a 2-second fail-fast test to make sure that
the postmaster.pid file was written to disk.

-- However, in extreme loading conditions, the underlying OS may be too busy
to accomplish that within 2 seconds. Removing the fast-fail means
that in that extreme loading case, pg_ctl will still be able to start,
assuming that the OS can write a file within 60 seconds.

-- This does mean that in cases where the pidfile cannot be written at
all-- for instance, if the disk is full-- the failure will take the full
60 seconds to be reported.