Fix CCL artifact names for sles11 builds [#132766731]
Previously, for sles11, MPP_ARCH fell to the default value of BLD_ARCH,
which was sles11_x86_64 (with an underscore). This meant the CCL builds
produced .zip files with underscores in the name. However, this file got
renamed in the pipeline via the `versioned_file:` property. The `put`
step used the `versioned_file` as the output name, but takes a `file`
parameter as an input name.

To fix, this change adds a setting for
`sles11_x86_64_MPP_ARCH=SLES11-x86_64` and fixes the put step `file`
parameters to match. The file name on s3 is preserved.

Signed-off-by: David Sharp <dsharp@pivotal.io>