Make queryfinish pending test more stable.
Add order by clause to test_and_limit query so that the ordering of the rows is
consistent. Without the order by clause, the limit clause will cause the query
to return the first 'n' rows which may be different for each run of the query.
Add an order by clause to dumpsort query to make the order of the rows
consistent and remove one of the cte in the join to make the test a bit faster.