Use config.h from ORCA to determine GPOS_DEBUG and arch flags.
Starting with ORCA version 2.2, there's a gpos/config.h file that contains
flags describing the compile-time options used to build the ORCA library.
Those flags affect binary compatibility, so it's important that e.g. if
ORCA was built with GPOS_DEBUG, the code that uses it (src/backend/gpopt
in this case) is also built with GPOS_DEBUG. Use the new gpos/config.h
for that, instead of deriving them ourselves and hoping that we reach the
same conclusions as whoever built ORCA.

This requires ORCA v2.2, so update releng.mk to download that.

(cherry picked from commit 60449fa24910a0007546e75d19eecee42de6471d)