Remove redundant check on bitmapindex
During IndexBuildHeapScan, there is extra check on whether the index is a
bitmap index. This was introduced back in 3.3, and no longer apply now. The
scenarios are covered in
src/test/tinc/tincrepo/mpp/gpdb/tests/storage/vacuum/reindex.

We remove this redundant check to make the code more consistent with GPDB5 and
upstream behaviors.

This is GPDB5 commit:
	commit 5deb836e9c3c7a24d26c1586b7666df918ca0f7c
	Author: Abhijit Subramanya <asubramanya@pivotal.io>
	Date:   Thu Jan 26 11:38:01 2017 -0800

	    Taking upstream behavior on IndexBuildHeapScan

	Signed-off-by: Xin Zhang <xzhang@pivotal.io>