Fix the calculation of hidden tuple ratio for Appendonly tables.
The function `AppendOnlyCompaction_GetHideRatio()` incorrectly truncated the
fractional part while calculating the ratio of hidden tuples which caused the
file to not be compacted (for e.g if the calculated threshold was 10.1% and the
guc was set to 10%, the segfile would not be compacted). This patch correctly
rounds up the calculated value to fix this issue.