Fix the new workfile regression tests to work with ORCA.
The query in hashjoin_spill was choosing a different plan with ORCA, where
the "other" side of the join was hashed than expected, and therefore it did
not spill to disk as expected. Turn the join the query into an OUTER JOIN,
to force the join order.

In materialize_spill and sis_mat_sort tests, the test queries are tailored
to produce a particular plan with a Materialize node that spills. ORCA
chose a different plan that didn't do that. I was not able to quickly
formulate the query in a way that would produce the same result with and
without ORCA, so just force ORCA off for the test. The point of the test
is to test workfile behavior at runtime, in executor, so that should be OK.

Likewise, the sisc_sort_spill test was producing different plan with ORCA,
without a Sort, so it wasn't testing what it was supposed to test. Disable
ORCA for sisc_sort_spilll too.

The helper function in sisc_mat_sort tests were producing "insufficient
memory reserved for statement" errors with ORCA. Not sure why, but I bumped
statement_mem slightly to not hit that limit anymore. It still spills to
disk, which is the point.

This is ported from
https://github.com/greenplum-db/gpdb/commit/5deeb4d30ab6d9439f5017cd5ee9e3acd3a638af.