Update TINC discover -s to replace '/' with '.' if -t is given
TINC discover would try to import the file instead of the class when
given something like this:
tinc.py discover -t tincrepo/mpp/gpdb/tests -s dispatch/queryfinish

The -t path is added to sys.path and the -s is imported. The -s needs
to be in dots for proper class importing.