Handle corrupted statslot [#137924345]
Use the actual element type of the array it's disassembling, rather than
trusting the type OID passed in by its caller.  This is needed because
sometimes the planner passes in a type OID that's only binary-compatible
with the target column's type, rather than being an exact match.

Ported the fix from postgres:
https://git.postgresql.org/gitweb/?p=postgresql.git;a=commitdiff;h=8c21b4e9226df1f6f16091e557fb313f5d308cde
Signed-off-by: Dhanashree Kashid <dhanashreek89@gmail.com>