Create database commands must use template0 as template
This applies to tests that create a database with checkpoint skip fault
injected.  Create database now requests checkpoint and if it is skipped, any
modifications to the template database pending to be written are potentially
lost.  The database template0 is guaranteed to not have any modifications
because user connections are not allowed on it.

Signed-off-by: Xin Zhang <xzhang@pivotal.io>