gpcloud: hash and sign the payload
```
When you send a request, you must tell Amazon S3 which of the preceding
options you have chosen in your signature calculation, by adding the
x-amz-content-sha256 header with one of the following values:

If you choose chunked upload options, set the header value to
STREAMING-AWS4-HMAC-SHA256-PAYLOAD.
If you choose to upload payload in a single chunk, set the header value
to the payload checksum (signed payload option), or set the value to the
literal string UNSIGNED-PAYLOAD (unsigned payload option).
```

ref:
http://docs.aws.amazon.com/AmazonS3/latest/API/sigv4-auth-using-authorization-header.html

We used to not hash or sign the payload, it's not quite secure with
HTTPS off, which is enabled by default although. And some compatible S3
services don't support "unsigned payload option", this commit fixes it.

Signed-off-by: Adam Lee <ali@pivotal.io>
Signed-off-by: Yuan Zhao <yuzhao@pivotal.io>